Modifications to allow multiple server configurations.
Originally had the stuff under the tcltk subdirectory; duplicated that
within a directory called 'default' and made changes to site-config,
tcltk/config.sh, default/config.sh to support this. The web startup
script was modified to default to the default server or, if SITENAME was
already set to some string or there was a $2 argument passed to the web
script, it would use that string to source from another server named
directory. This means:

  web start tcltk

Would start up the stuff in the tcltk subdirectory.