Modified db and web to be able to use external site configs.
Each site will be in it's own directory at the same level as this one
(i.e. a sibling of the top level directory of this distribution). This
is because each site will be in its own separate git repository.

To use, simply pass the name to the db or web script as the third
argument.